requirements.txt
aiogram==3.7.0
asyncpg==0.29.0
python-dotenv==1.0.0
fastapi==0.111.0
uvicorn==0.30.0